WIP: Introduce ENV var to specify CPI location
`OPENSTACK_CPI_BIN` can be set to an existing CPI binary.
CPI installation will be skipped in that case.

WIP because of two variables in Context class.
Check for valid env var could happen in Context.new
Additionally, it is not checked if context.cpi_bin_path is
used in execute_specs

[#131970415](https://www.pivotaltracker.com/story/show/131970415)

Signed-off-by: Felix Riegger <felix.riegger@sap.com>
Signed-off-by: Tom Kiemes <tom.kiemes@sap.com>
Signed-off-by: Felix Riegger <felix.riegger@sap.com>